Make whole_essay work by regrouping document segments
Move whole_essay a model where each batch is formed of segments which
are tagged with their document ID. This doesn't play nicely with Captum,
so in this case the segmentation is peformed after Captum rather than
before. TODO: Investigate/discuss whether this whole thing should've
been put inside the model in the first place.